float sample to ticks

+++ include/tpglibs/TPGenerator.hpp
@@ -32,7 +32,7 @@ class TPGenerator {
   static const uint8_t m_num_channels_per_pipeline = 16; // AVX2 with int16 data samples allows us to process 16 channels.
   uint8_t m_num_pipelines = 0;  // Gets set inside configure.
   std::vector<AVXPipeline> m_tpg_pipelines;
-  int m_sample_tick_difference;
+  float m_sample_tick_difference;
   std::vector<uint16_t> m_sot_minima{1,1,1};  // Defaults to 1 for all planes.
   std::shared_ptr<ProcessorMetricCollector<__m256i>> m_processor_metric_collector_ptr {nullptr};
   bool m_tpg_metric_collect_enabled {false};
@@ -47,7 +47,7 @@ class TPGenerator {
      */
     void configure(const std::vector<std::pair<std::string, nlohmann::json>>& configs,
                    const std::vector<std::pair<dunedaq::trgdataformats::channel_t, int16_t>> channel_plane_numbers,
-                   const int sample_tick_difference);
+                   const float sample_tick_difference);
 
     /**
      * @brief Set the minimum samples over threshold for a TP according to plane.
@@ -126,7 +126,7 @@ class TPGenerator {
           std::vector<dunedaq::trgdataformats::TriggerPrimitive> tps = m_tpg_pipelines[p].process(expanded_subframe);
 
           for (auto tp : tps) {
-            tp.time_start = (t - tp.samples_over_threshold) * m_sample_tick_difference + timestamp;
+            tp.time_start = static_cast<int64_t>((t - tp.samples_over_threshold) * m_sample_tick_difference) + timestamp;
             tp_aggr.push_back(tp);
           }
           cursor += register_alignment / 8; // Numerator is in bits. Need bytes.
